shadfxml/ififuncs - better error handling and windows support

+++ shadfxml.py
@@ -36,46 +36,51 @@ def main(args_):
     user = ififuncs.get_user()
     for root, _, _ in os.walk(source):
         if ififuncs.check_for_uuid_generic(root) is not False:
-            uuid_path = root
-            uuid = os.path.basename(uuid_path)
+            print(" - Processing %s" % root)
+            oe_path = root
+            uuid = ififuncs.check_for_uuid_generic(root)
+            uuid_path = os.path.join(oe_path, uuid)
             logs_dir = os.path.join(uuid_path, 'logs')
             metadata_dir = os.path.join(uuid_path, 'metadata')
             dfxml = os.path.join(metadata_dir, uuid + '_dfxml.xml')
             new_log_textfile = os.path.join(logs_dir, uuid) + '_sip_log.log'
-            new_manifest_textfile = os.path.join(os.path.dirname(uuid_path), uuid) + '_manifest.md5'
-            ififuncs.generate_log(
-                new_log_textfile,
-                'EVENT = shadfxml.py started'
-            )
-            ififuncs.generate_log(
-                new_log_textfile,
-                'eventDetail=shadfxml.py %s' % ififuncs.get_script_version('shadfxml.py')
-            )
-            ififuncs.generate_log(
-                new_log_textfile,
-                'Command line arguments: %s' % args
-            )
-            ififuncs.generate_log(
-                new_log_textfile,
-                'EVENT = agentName=%s' % user
-            )
-            if not os.path.isfile(dfxml):
-                print('Generating Digital Forensics XML')
-                dfxml = accession.make_dfxml(args, uuid_path, uuid)
-                ififuncs.generate_log(
-                    new_log_textfile,
-                    'EVENT = Metadata extraction - eventDetail=File system metadata extraction using Digital Forensics XML, eventOutcome=%s, agentName=makedfxml' % (dfxml)
-                )
-                ififuncs.manifest_update(new_manifest_textfile, dfxml)
             sha512_manifest = os.path.join(
                 os.path.dirname(uuid_path), uuid + '_manifest-sha512.txt'
             )
-            if not os.path.isfile(sha512_manifest):
-                sha512_log = manifest.main([uuid_path, '-sha512', '-s'])
-                ififuncs.merge_logs_append(sha512_log, new_log_textfile, new_manifest_textfile)
-                ififuncs.checksum_replace(sha512_manifest, new_log_textfile, 'sha512')
-                os.remove(sha512_log)
-            ififuncs.sort_manifest(new_manifest_textfile)
+            if not os.path.isfile(dfxml) and os.path.isfile(sha512_manifest):
+                new_manifest_textfile = os.path.join(os.path.dirname(uuid_path), uuid) + '_manifest.md5'
+                ififuncs.generate_log(
+                    new_log_textfile,
+                    'EVENT = shadfxml.py started'
+                )
+                ififuncs.generate_log(
+                    new_log_textfile,
+                    'eventDetail=shadfxml.py %s' % ififuncs.get_script_version('shadfxml.py')
+                )
+                ififuncs.generate_log(
+                    new_log_textfile,
+                    'Command line arguments: %s' % args
+                )
+                ififuncs.generate_log(
+                    new_log_textfile,
+                    'EVENT = agentName=%s' % user
+                )
+                if not os.path.isfile(dfxml):
+                    print('Generating Digital Forensics XML')
+                    dfxml = accession.make_dfxml(args, uuid_path, uuid)
+                    ififuncs.generate_log(
+                        new_log_textfile,
+                        'EVENT = Metadata extraction - eventDetail=File system metadata extraction using Digital Forensics XML, eventOutcome=%s, agentName=makedfxml' % (dfxml)
+                    )
+                    ififuncs.manifest_update(new_manifest_textfile, dfxml)
+                if not os.path.isfile(sha512_manifest):
+                    sha512_log = manifest.main([uuid_path, '-sha512', '-s'])
+                    ififuncs.merge_logs_append(sha512_log, new_log_textfile, new_manifest_textfile)
+                    ififuncs.checksum_replace(sha512_manifest, new_log_textfile, 'sha512')
+                    os.remove(sha512_log)
+                ififuncs.sort_manifest(new_manifest_textfile)
+            else:
+                print("Exiting as this package already has DFXML and SHA512")
 
 if __name__ == '__main__':
     main(sys.argv[1:])
